refactor to remove duplication
