+++ app/validators/address/international.validator.js
@@ -8,7 +8,7 @@
 
 const Joi = require('joi')
 const { countries } = require('../../lib/static-lookups.lib.js')
-const { invalidStartCharacters } = require('../helpers/notify-address-line.validator.js')
+const { addressLineValidator } = require('./addressLine.validator.js')
 
 /**
  * Validates data submitted for the `address/{sessionId}/international` page
@@ -20,19 +20,7 @@ const { invalidStartCharacters } = require('../helpers/notify-address-line.valid
  */
 function go(payload) {
   const schema = Joi.object({
-    addressLine1: Joi.string().required().custom(_addressLineCustom).messages({
-      'any.required': 'Enter address line 1',
-      'string.custom': 'Address line 1 cannont start with a special character'
-    }),
-    addressLine2: Joi.string().optional().custom(_addressLineCustom).messages({
-      'string.custom': 'Address line 2 cannont start with a special character'
-    }),
-    addressLine3: Joi.string().optional().custom(_addressLineCustom).messages({
-      'string.custom': 'Address line 3 cannont start with a special character'
-    }),
-    addressLine4: Joi.string().optional().custom(_addressLineCustom).messages({
-      'string.custom': 'Address line 4 cannont start with a special character'
-    }),
+    ...addressLineValidator(),
     country: Joi.string()
       .required()
       .valid(...countries)
@@ -43,13 +31,6 @@ function go(payload) {
   })
 
   return schema.validate(payload, { abortEarly: false })
-}
-
-function _addressLineCustom(value, helper) {
-  if (invalidStartCharacters(value)) {
-    return helper.error('string.custom')
-  }
-  return value
 }
 
 module.exports = {
