+++ app/validators/address/manual.validator.js
@@ -7,7 +7,7 @@
  */
 
 const { postcodeValidator } = require('postcode-validator')
-const { invalidStartCharacters } = require('../helpers/notify-address-line.validator.js')
+const { addressLineValidator } = require('./addressLine.validator.js')
 
 const Joi = require('joi')
 
@@ -21,19 +21,7 @@ const Joi = require('joi')
  */
 function go(payload) {
   const schema = Joi.object({
-    addressLine1: Joi.string().required().custom(_addressLineCustom).messages({
-      'any.required': 'Enter address line 1',
-      'string.custom': 'Address line 1 cannont start with a special character'
-    }),
-    addressLine2: Joi.string().optional().custom(_addressLineCustom).messages({
-      'string.custom': 'Address line 2 cannont start with a special character'
-    }),
-    addressLine3: Joi.string().optional().custom(_addressLineCustom).messages({
-      'string.custom': 'Address line 3 cannont start with a special character'
-    }),
-    addressLine4: Joi.string().optional().custom(_addressLineCustom).messages({
-      'string.custom': 'Address line 4 cannont start with a special character'
-    }),
+    ...addressLineValidator(),
     postcode: Joi.string()
       .required()
       .custom((value, helper) => {
@@ -49,13 +37,6 @@ function go(payload) {
   })
 
   return schema.validate(payload, { abortEarly: false })
-}
-
-function _addressLineCustom(value, helper) {
-  if (invalidStartCharacters(value)) {
-    return helper.error('string.custom')
-  }
-  return value
 }
 
 module.exports = {
